backend/requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
groq==0.9.0
python-dotenv==1.0.0
asyncpg==0.29.0
